AppRouter.js: - import statement adjusted
GameLobby.js:
- file added

GameStart.js:
- new file replacing GameLobby.js
